Refactor main method in GameOfLife project
The main method has been refactored in the GameOfLife project. Specific classes were imported instead of importing the entire swing and awt packages, and model, view, and controller instances were changed to use factory methods. Modeling is done with a createModel() method, the view now uses a noarg constructor, and the controller now uses a method that includes a random initialization.